+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ## 0.15.4
+* Mono-debug now uses the terminal service provided by VS Code. You can now use the launch config attribute `console` for selecting one of `internalConsole`, `integratedTerminal`, `externalTerminal`. The 'integratedTerminal' and 'externalTerminal' can be further configured through workspace or user settings.
 * Added support for configuring handled and unhandled exceptions. Use the "Debug: Configure Exceptions" command.
 
 ## 0.15.3